replace spread operator with mutation
reason: APL performance was *pretty bad* in production mode but not
development mode. it turns out that in prod this gets compiled down to
an es5-compatible method call that is way less efficient than modifying
the object in place.

this makes things faster in dev mode by about 50ms on my sylvanas test
log (2.6x). if it were under 100ms in prod i probably wouldn't care, but
it was over 400ms and the majority of the time was spent doing the
spread operation. we'll see how this performs in production and go from there.